requirements.txt
langchain
openai
python-dotenv
anthropic
flask
flask-cors
chromadb
tiktoken
authlib